refactor: merge SfTable components to vue file
