--- packages/vue/src/components/organisms/SfTable/SfTable.html
@@ -1,3 +1,0 @@
-<table class="sf-table">
-  <slot />
-</table>
