+++ packages/vue/src/components/organisms/SfTable/SfTable.vue
@@ -1,5 +1,37 @@
-<script src="./SfTable.js"></script>
-<template lang="html" src="./SfTable.html"></template>
+<template>
+  <table class="sf-table">
+    <slot />
+  </table>
+</template>
+<script>
+import Vue from "vue";
+import SfTableRow from "./_internal/SfTableRow.vue";
+import SfTableData from "./_internal/SfTableData.vue";
+import SfTableHeader from "./_internal/SfTableHeader.vue";
+import SfTableHeading from "./_internal/SfTableHeading.vue";
+Vue.component("SfTableRow", SfTableRow);
+Vue.component("SfTableData", SfTableData);
+Vue.component("SfTableHeader", SfTableHeader);
+Vue.component("SfTableHeading", SfTableHeading);
+export default {
+  name: "SfTable",
+  provide() {
+    const table = {};
+    Object.defineProperty(table, "updateColumnsCount", {
+      value: this.updateColumnsCount
+    });
+    return { table };
+  },
+  methods: {
+    updateColumnsCount(columnsCount) {
+      this.$el.style.setProperty(
+        "--mobile-column",
+        Math.ceil(columnsCount / 2)
+      );
+    }
+  }
+};
+</script>
 <style lang="scss">
 @import "~@storefront-ui/shared/styles/components/SfTable.scss";
 </style>
