--- packages/vue/src/components/organisms/SfTable/_internal/SfTableData.html
@@ -1,3 +1,0 @@
-<td class="sf-table__data">
-    <slot/>
-</td>
