--- packages/vue/src/components/organisms/SfTable/_internal/SfTableData.js
@@ -1,3 +1,0 @@
-export default {
-  name: "SfTableData"
-};
