+++ packages/vue/src/components/organisms/SfTable/_internal/SfTableData.vue
@@ -1,2 +1,10 @@
-<script src="./SfTableData.js"></script>
-<template lang="html" src="./SfTableData.html"></template>
+<template>
+  <td class="sf-table__data">
+    <slot />
+  </td>
+</template>
+<script>
+export default {
+  name: "SfTableData"
+};
+</script>
