--- packages/vue/src/components/organisms/SfTable/_internal/SfTableHeader.html
@@ -1,3 +1,0 @@
-<th class="sf-table__header">
-    <slot/>
-</th>
