--- packages/vue/src/components/organisms/SfTable/_internal/SfTableHeader.js
@@ -1,3 +1,0 @@
-export default {
-  name: "SfTableHeader"
-};
