+++ packages/vue/src/components/organisms/SfTable/_internal/SfTableHeader.vue
@@ -1,2 +1,10 @@
-<script src="./SfTableHeader.js"></script>
-<template lang="html" src="./SfTableHeader.html"></template>
+<template>
+  <th class="sf-table__header">
+    <slot />
+  </th>
+</template>
+<script>
+export default {
+  name: "SfTableHeader"
+};
+</script>
