--- packages/vue/src/components/organisms/SfTable/_internal/SfTableHeading.html
@@ -1,4 +1,0 @@
-<tr class="sf-table__heading">
-    <slot/>
-</tr>
-
