--- packages/vue/src/components/organisms/SfTable/_internal/SfTableHeading.js
@@ -1,10 +1,0 @@
-export default {
-  name: "SfTableHeading",
-  inject: ["table"],
-  mounted() {
-    if (!this.$slots.default) return;
-    this.table.updateColumnsCount(
-      this.$slots.default.filter(node => node.tag).length
-    );
-  }
-};
