+++ packages/vue/src/components/organisms/SfTable/_internal/SfTableHeading.vue
@@ -1,2 +1,17 @@
-<script src="./SfTableHeading.js"></script>
-<template lang="html" src="./SfTableHeading.html"></template>
+<template>
+  <tr class="sf-table__heading">
+    <slot />
+  </tr>
+</template>
+<script>
+export default {
+  name: "SfTableHeading",
+  inject: ["table"],
+  mounted() {
+    if (!this.$slots.default) return;
+    this.table.updateColumnsCount(
+      this.$slots.default.filter(node => node.tag).length
+    );
+  }
+};
+</script>
