--- packages/vue/src/components/organisms/SfTable/_internal/SfTableRow.html
@@ -1,3 +1,0 @@
-<tr class="sf-table__row">
-  <slot/>
-</tr>
