--- packages/vue/src/components/organisms/SfTable/_internal/SfTableRow.js
@@ -1,10 +1,0 @@
-export default {
-  name: "SfTableRow",
-  inject: ["table"],
-  mounted() {
-    if (!this.$slots.default) return;
-    this.table.updateColumnsCount(
-      this.$slots.default.filter(node => node.tag).length
-    );
-  }
-};
