+++ packages/vue/src/components/organisms/SfTable/_internal/SfTableRow.vue
@@ -1,2 +1,17 @@
-<script src="./SfTableRow.js"></script>
-<template lang="html" src="./SfTableRow.html"></template>
+<template>
+  <tr class="sf-table__row">
+    <slot />
+  </tr>
+</template>
+<script src="./SfTableRow.js">
+export default {
+  name: "SfTableRow",
+  inject: ["table"],
+  mounted() {
+    if (!this.$slots.default) return;
+    this.table.updateColumnsCount(
+      this.$slots.default.filter(node => node.tag).length
+    );
+  }
+};
+</script>
